// src/utilities/enum.ts
export const enum TYPE_ACTION {
  ADD_TODO = 'AGREGAR',
  TODO_COMPLETED = 'TODO_COMPLETED',
  DELETE_TODO = 'DELETE_TODO',
  CLEAR_COMPLETED_TODOS = 'CLEAR_COMPLETED_TODO',
  REORDER_TODOS = 'REORDER_TODOS',
  FILTER_ALL = 'DEFAULT',
  FILTER_ACTIVE = 'FILTER_ACTIVE',
  FILTER_COMPLETED = 'FILTER_COMPLETED',
}